compelete-homepage without styling
