+++ components/Homepage.js
@@ -1,12 +1,14 @@
 import React from 'react';
 import { View, StyleSheet } from 'react-native';
 import HomepageCards from './HomepageCards';
+import MapComponent from './MapComponent';
 
 export default function Homepage() {
   return (
     <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
       <View style={styles.HomeBtmSection}>
-      <HomepageCards  />
+      <MapComponent />
+      <HomepageCards />
       </View>
     </View>
   );
